Create InTheAreaOriginal.cmp
Code Highlights:
Lightning Component markup consists of a mixture of XML tags and standard HTML tags. We’ll add styling for the component later.
The component uses flexipage:availableForAllPageTypes to enable the component to be used in Lightning App Builder.
The component also uses force:appHostable to allow the component to be used as a navigation element in Lightning Experience or the Salesforce1 app.
flexipage:availableForRecordHome and force:hasRecordId make the component available for use on a Record home page in Lightning Experience.
Several elements have aura:id applied to allow us to find the elements from our controller logic.