Code Highlights
All class names in the Salesforce Lightning Design System are prepended with slds-.
The Salesforce Lightning Design System uses BEM naming, which stands for Block, Element, Modifier. The block is the first part of the class name, such as slds-form-element. The element is identified with a double dash, such as slds-text-heading--small, and the modifier is identified by a double underscore, such as slds-list__item. While this might appear more verbose, it makes it easy for us to quickly identify and understand the styling of the building blocks of our component.