Create IntheArea.cmp
Code Highlights
A Lightning Component has a built-in event that can trigger a handler named init when the component loads. The action defines the name of the function in the component’s controller that is called when the handler is fired. In this case, we’re looking for a function named doInit.